requirements.txt
Django==2.2
django-cors-headers==3.5.0
django-hosts==4.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
